requirements.txt
crewai
crewai-tools
docker
python-dotenv
litellm
pydantic
fastapi
uvicorn
chromadb
websockets
